% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_class.R
\name{classify_commits_in_db}
\alias{classify_commits_in_db}
\title{Классификация коммитов в базе данных}
\usage{
classify_commits_in_db(
  conn,
  model_obj,
  table_name = "git_commit_history",
  author_name = NULL,
  batch_size = 1000,
  repo_id = NULL,
  group_mode = c("en", "ru")
)
}
\description{
Классификация коммитов в базе данных
}
