% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL.R
\name{clone_or_pull}
\alias{clone_or_pull}
\title{Clone or update a Git repository}
\usage{
clone_or_pull(mode, repo_url = NULL, clone_dir = NULL, local_path = NULL)
}
\arguments{
\item{mode}{0 = local, 1 = remote}

\item{repo_url}{GitHub URL (for mode = 1)}

\item{clone_dir}{Where to clone (for mode = 1)}

\item{local_path}{Path to local repo (for mode = 0)}
}
\value{
Path to repository
}
\description{
Clone or update a Git repository
}
