% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL.R
\name{db_summary}
\alias{db_summary}
\title{Get a quick summary of the database}
\usage{
db_summary(con = NULL, db_path = "git.duckdb")
}
\arguments{
\item{con}{Database connection (if NULL, a temporary connection is created)}

\item{db_path}{Path to the database (used only if con is NULL)}
}
\value{
List with summary information
}
\description{
Get a quick summary of the database
}
\examples{
\dontrun{
summary <- db_summary()
print(summary$total_commits)
print(summary$date_range)
}

}
