% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL.R
\name{delete_repository}
\alias{delete_repository}
\title{Delete a repository from the database}
\usage{
delete_repository(repo_name)
}
\arguments{
\item{repo_name}{Name of the repository to delete}
}
\value{
Invisibly returns TRUE if deleted, FALSE if not found
}
\description{
Removes all data associated with a repository from all tables:
\itemize{
\item git_file_changes
\item git_commit_history
\item repo_metadata
\item repo_path
}
}
\examples{
\dontrun{
delete_repository("dbipAnalyzer")
}

}
