% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_class.R
\name{evaluate_classifier}
\alias{evaluate_classifier}
\title{Оценка на тестовой выборке (с удалением стоп-слов)}
\usage{
evaluate_classifier(csv_path = NULL, train_ratio = 0.8, max_features = 1000)
}
\description{
Оценка на тестовой выборке (с удалением стоп-слов)
}
