% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalies_detection.R
\name{format_anomalies_for_hr}
\alias{format_anomalies_for_hr}
\title{HR‑форматирование аномалий (без параметра period)}
\usage{
format_anomalies_for_hr(
  rule_anomalies,
  ml_anomalies = NULL,
  frequent_edits = NULL
)
}
\description{
HR‑форматирование аномалий (без параметра period)
}
