% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL.R
\name{get_commits}
\alias{get_commits}
\title{Parse git log -p output into commit blocks}
\usage{
get_commits(repo_path)
}
\arguments{
\item{repo_path}{Path to local Git repository}
}
\value{
List of character vectors, each vector is one commit block
}
\description{
Parse git log -p output into commit blocks
}
