% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anomalies_detection.R
\name{get_frequent_file_edits}
\alias{get_frequent_file_edits}
\title{Частые правки одного файла (>10 раз в день)}
\usage{
get_frequent_file_edits(
  conn,
  username = NULL,
  since = NULL,
  until = NULL,
  threshold = 10
)
}
\description{
Частые правки одного файла (>10 раз в день)
}
