% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_analysis.R
\name{get_ml_anomalies}
\alias{get_ml_anomalies}
\title{Получение ML-аномалий (Isolation Forest)}
\usage{
get_ml_anomalies(
  conn,
  author_name = NULL,
  threshold = 0.95,
  return_features = TRUE,
  since = NULL,
  until = NULL
)
}
\description{
Получение ML-аномалий (Isolation Forest)
}
