% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL.R
\name{parse_hunk_line}
\alias{parse_hunk_line}
\title{Parse a hunk line (@ ... @) to extract line numbers
Supports multiple formats:
\itemize{
\item @ -10,5 +10,8 @
\item @ -10 +10 @
\item @ -10,5 +10 @
\item @ -10 +10,5 @
}}
\usage{
parse_hunk_line(line)
}
\arguments{
\item{line}{A string starting with "@ "}
}
\value{
Numeric vector of length 4: c(start_del, count_del, start_add, count_add) or NULL if parsing fails
}
\description{
Parse a hunk line (@ ... @) to extract line numbers
Supports multiple formats:
\itemize{
\item @ -10,5 +10,8 @
\item @ -10 +10 @
\item @ -10,5 +10 @
\item @ -10 +10,5 @
}
}
