% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETL.R
\name{reset_db}
\alias{reset_db}
\title{Reset the DuckDB database by deleting the database file and its WAL}
\usage{
reset_db(db_path = "git.duckdb")
}
\arguments{
\item{db_path}{Path to the DuckDB database file (default: "git.duckdb")}
}
\value{
Invisibly returns TRUE
}
\description{
Reset the DuckDB database by deleting the database file and its WAL
}
\examples{
\dontrun{
reset_db("git.duckdb")
}

}
