% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit_class.R
\name{train_commit_classifier}
\alias{train_commit_classifier}
\title{Обучение модели с удалением стоп-слов и отбором top max_features признаков}
\usage{
train_commit_classifier(csv_path = NULL, max_features = 1000)
}
\description{
Обучение модели с удалением стоп-слов и отбором top max_features признаков
}
